[render] added window focus event; added FreeCamera class in construct; fixes for some math functions; float instead of double for dt; misc changes

+++ src/Construct/Construct/Construct.h
@@ -8,7 +8,7 @@ namespace construct
 	{
 	public:
 		Construct();
-		void Update(double dt) override;
+		void Update(float dt) override;
 	private:
 		void CreateTestMesh();
 	};
